video-to-process-doc-local (1)/video-to-process-doc/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
pydantic==2.6.1
python-docx==1.1.0
openai==1.23.6
azure-cognitiveservices-speech==1.37.0
tiktoken==0.7.0
ffmpeg-python==0.2.0
